refactor(sdk): split experiment.ts into focused modules
Extract types, evaluation sink, arm dispatch, shadow comparison,
validation, and runtime primitives into experiment-*.ts siblings;
experiment.ts keeps the orchestrator (1099 -> 260 lines). No behavior
change.
